refactor: split publisher for transfer process and contract negotiation
