+++ extensions/control-plane/tasks/lib/tasks-nats-lib/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
