--- extensions/control-plane/tasks/publisher/tasks-publisher-nats/src/main/java/org/eclipse/edc/virtual/controlplane/events/publisher/nats/NatsTaskPublisherExtension.java
@@ -1,70 +1,0 @@
-/*
- *  Copyright (c) 2026 Metaform Systems, Inc.
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Metaform Systems, Inc. - initial API and implementation
- *
- */
-
-package org.eclipse.edc.virtual.controlplane.events.publisher.nats;
-
-import io.nats.client.Nats;
-import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
-import org.eclipse.edc.runtime.metamodel.annotation.Inject;
-import org.eclipse.edc.runtime.metamodel.annotation.Setting;
-import org.eclipse.edc.runtime.metamodel.annotation.Settings;
-import org.eclipse.edc.spi.monitor.Monitor;
-import org.eclipse.edc.spi.system.ServiceExtension;
-import org.eclipse.edc.spi.system.ServiceExtensionContext;
-import org.eclipse.edc.spi.types.TypeManager;
-import org.eclipse.edc.virtual.controlplane.tasks.TaskObservable;
-
-import java.time.Clock;
-
-public class NatsTaskPublisherExtension implements ServiceExtension {
-
-
-    @Inject
-    private TypeManager typeManager;
-
-    @Configuration
-    private NatsPublisherConfig natsPublisherConfig;
-
-
-    @Inject
-    private Clock clock;
-
-    @Inject
-    private Monitor monitor;
-
-    @Inject
-    private TaskObservable taskObservable;
-
-    @Override
-    public void initialize(ServiceExtensionContext context) {
-        try {
-            var connection = Nats.connect(natsPublisherConfig.url());
-            var js = connection.jetStream();
-            var publisher = new NatsTaskPublisher(js, monitor, () -> typeManager.getMapper());
-
-            taskObservable.registerListener(publisher);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Settings
-    public record NatsPublisherConfig(
-            @Setting(key = "edc.nats.cn.publisher.url", description = "The URL of the NATS server to connect to for publishing contract negotiation events.", defaultValue = "nats://localhost:4222")
-            String url,
-            @Setting(key = "edc.nats.cn.publisher.subject-prefix", description = "The prefix for the subjects", defaultValue = "negotiations")
-            String subjectPrefix
-    ) {
-    }
-}
