+++ settings.gradle.kts
@@ -54,11 +54,12 @@ include(":extensions:cel:cel-extension")
 include(":extensions:cel:cel-store-sql")
 include(":extensions:control-plane:tasks:listener:tasks-store-listener")
 include(":extensions:control-plane:tasks:listener:tasks-store-poll-executor")
-include(":extensions:control-plane:tasks:publisher:tasks-publisher-nats")
+include(":extensions:control-plane:tasks:publisher:negotiation-tasks-publisher-nats")
+include(":extensions:control-plane:tasks:publisher:transfer-tasks-publisher-nats")
 include(":extensions:control-plane:tasks:store:tasks-store-sql")
 include(":extensions:control-plane:tasks:subscriber:negotiation-tasks-subscriber-nats")
 include(":extensions:control-plane:tasks:subscriber:transfer-tasks-subscriber-nats")
-include(":extensions:control-plane:tasks:lib:tasks-subscriber-nats-lib")
+include(":extensions:control-plane:tasks:lib:tasks-nats-lib")
 
 include(":extensions:lib:nats-lib")
 include(":extensions:common:api:api-authentication")
